TIP-840: move Currency entity
