+++ src/Pim/Bundle/EnrichBundle/Controller/CurrencyController.php
@@ -2,9 +2,9 @@
 
 namespace Pim\Bundle\EnrichBundle\Controller;
 
+use Akeneo\Channel\Component\Model\Currency;
 use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
-use Pim\Bundle\CatalogBundle\Entity\Currency;
 use Pim\Bundle\EnrichBundle\Flash\Message;
 use Pim\Component\Catalog\Exception\LinkedChannelException;
 use Symfony\Component\HttpFoundation\JsonResponse;
